dispatch: the interface lifecycle and the egress path in their own modules
reconcile rebuilds every stale interface before the dispatcher evicts and
notifies per rebuilt one; with two stale interfaces the second's rebind now
precedes the first's eviction, and an interface's "recovery complete" line
and counter now precede its own evictions.
